max-statistic for a 2x3 table

Compute pairwise linkage disequilibrium between genetic markers

Usage

LD(g1, ...)

# S3 method for snp
LD(g1, g2, ...)

# S3 method for setupSNP
LD(g1, SNPs, ...)

LDplot(x, digits = 3, marker, distance, which = c("D", "D'",
    "r", "X^2", "P-value", "n", " "), ...)

LDtable(x, colorcut = c(0, 0.01, 0.025, 0.05, 0.1, 1), 
       colors = heat.colors(length(colorcut)),
       textcol = "black", digits = 3, show.all = FALSE, 
       which = c("D", "D'", "r", "X^2", "P-value", "n"), 
       colorize = "P-value", cex, ...)

Arguments

g1

genotype object or dataframe containing genotype objects

g2

genotype object (ignored if g1 is a dataframe)

SNPs

columns containing SNPs

x

LD or LD.data.frame object

digits

Number of significant digits to display

which

Name(s) of LD information items to be displayed

colorcut

P-value cutoffs points for colorizing LDtable

colors

Colors for each P-value cutoff given in 'colorcut' for LDtable

textcol

Color for text labels for LDtable

marker

Marker used as 'comparator' on LDplot. If omitted separate lines for each marker will be displayed

distance

Marker location, used for locating of markers on LDplot.

show.all

If TRUE, show all rows/columns of matrix. Otherwise omit completely blank rows/columns.

colorize

LD parameter used for determining table cell colors

cex

Scaling factor for table text. If absent, text will be scaled to fit within the table cells.

...

Optional arguments ('plot.LD.data.frame' passes these to 'LDtable' and 'LDplot').

References

genetics R package by Gregory Warnes et al. (warnes@bst.rochester.edu)

Author

functions adapted from LD, LDtable and LDplot in package genetics by Gregory Warnes et al. (warnes@bst.rochester.edu)

Value

None

See also