Association analysis between a single SNP and a given phenotype

This function carries out an association analysis between a single SNP and a dependent variable (phenotype) under five different genetic models (inheritance patterns): codominant, dominant, recessive, overdominant and log-additive. The phenotype may be quantitative or categorical. In the second case (e.g. case-control studies) this variable must be of class 'factor' with two levels.

Usage

association(formula, data, model=c("all"), model.interaction=
       c("codominant"), subset, name.snp = NULL, quantitative = 
       is.quantitative(formula,data), genotypingRate= 0, 
       level = 0.95, ...)

Arguments

formula

a symbolic description of the model to be fited (a formula object). It might have either a continuous variable (quantitative traits) or a factor variable (case-control studies) as the response on the left of the ~ operator and a term corresponding to the SNP on the right. This term must be of class snp (e.g. ~snp(var), where var is a given SNP), and it is required. Terms with additional covariates on the right of the ~ operator may be added to fit an adjusted model (e.g., ~var1+var2+...+varN+SNP). The formula allows to incorporate more than one object of class snp. In that case, the analysis is done for the first SNP which appears in the formula adjusted by the others covariates and other additional SNPs.

data

a required dataframe of class 'setupSNP' containing the variables in the model.

model

a character string specifying the type of genetic model (mode of inheritance) for the SNP. This indicates how the genotypes should be collapsed. Possible values are "codominant", "dominant", "recessive", "overdominant", "additive" or "all". The default is "all" that fits the 5 possible genetic models. Only the first words are required, e.g "co", "do", etc.

model.interaction

a character string specifying the type of genetic model (mode of inheritance) assumed for the SNP when it is included in a interaction term. Possible values are "codominant", "dominant", "recessive", "overdominant". The default is "codominant".

subset

an optional vector specifying a subset of observations to be used in the fitting process

name.snp

optional label of the SNP variable to be printed.

quantitative

logical value indicating whether the phenotype (that which is in the left of the operator ~ in 'formula' argument) is quantitative. The function 'is.quantitative' returns FALSE when the phenotype is a variable with two categories (i.e. indicating case-control status). Thus, it is not a required argument but it may be modified by the user.

genotypingRate

minimum percentage of genotype rate for the SNP to be analyzed. Default is 0% (e.g. all SNPs are analyzed). This parameter should not be changed. It is used in the function 'WGassociation'.

level

signification level for confidence intervals.

...

Other arguments to be passed through glm function

Details

This function should be called by the user when we are interested in analyzing an unique SNP. It is recommended to use WGassociation function when more than one SNP is studied.

Value

For each genetic model (codominant, dominant, recessive, overdominant, and log-additive) the function gives a matrix with sample size and percentages for each genotype, the Odds Ratio and its 95% confidence interval (taking the most frequent homozygous genotype as the reference), the p-value corresponding to the likelihood ratio test obtained from a comparison with the null model, and the Akaike Information Criterion (AIC) of each genetic model. In the case of analyzing a quantitative trait, the function returns a matrix with sample size, mean and standard errors for each genotype, mean difference and its 95% confidence interval with respect to the most frequent homozygous genotype, the p-value obtained from an overall gene effect and the Akaike Information Criterion (AIC) of each genetic model.

When an interaction term (a categorical covariate with an SNP) is included in the model, three different tables are given. The first one correponds to the full interaction matrix where the ORs (or mean differences if a quantitative trait is analyzed) are expressed with respect to the non variant genotype and the first category of the covariate. The other two tables show the ORs and their 95% confidence intervals for both marginal models. P values for interaction and trend are also showed in the output.

References

JR Gonzalez, L Armengol, X Sole, E Guino, JM Mercader, X Estivill, V Moreno. SNPassoc: an R package to perform whole genome association studies. Bioinformatics, 2007;23(5):654-5.

Iniesta R, Guino E, Moreno V. Statistical analysis of genetic polymorphisms in epidemiological studies. Gac Sanit. 2005;19(4):333-41.

Elston RC. Introduction and overview. Statistical methods in genetic epidemiology. Stat Methods Med Res. 2000;9:527-41.

See also

Examples

data(SNPs)

# first, we create an object of class 'setupSNP'
datSNP<-setupSNP(SNPs,6:40,sep="")

# case-control study, crude analysis
association(casco~snp10001, data=datSNP)
#> 
#> SNP: snp10001  adjusted by: 
#>               0    %   1    %   OR lower upper p-value   AIC
#> Codominant                                                  
#> T/T          24 51.1  68 61.8 1.00             0.13234 193.6
#> C/T          21 44.7  32 29.1 0.54  0.26  1.11              
#> C/C           2  4.3  10  9.1 1.76  0.36  8.64              
#> Dominant                                                    
#> T/T          24 51.1  68 61.8 1.00             0.21184 194.1
#> C/T-C/C      23 48.9  42 38.2 0.64  0.32  1.28              
#> Recessive                                                   
#> T/T-C/T      45 95.7 100 90.9 1.00             0.27152 194.4
#> C/C           2  4.3  10  9.1 2.25  0.47 10.69              
#> Overdominant                                                
#> T/T-C/C      26 55.3  78 70.9 1.00             0.06129 192.1
#> C/T          21 44.7  32 29.1 0.51  0.25  1.03              
#> log-Additive                                                
#> 0,1,2        47 29.9 110 70.1 0.87  0.51  1.47 0.59448 195.4

# case-control study, adjusted by sex and arterial blood pressure
association(casco~sex+snp10001+blood.pre, data=datSNP)
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> 
#> SNP: snp10001  adjusted by: sex blood.pre 
#>               0    %   1    %   OR lower upper p-value   AIC
#> Codominant                                                  
#> T/T          24 51.1  68 61.8 1.00             0.15410 195.8
#> C/T          21 44.7  32 29.1 0.55  0.26  1.14              
#> C/C           2  4.3  10  9.1 1.74  0.35  8.63              
#> Dominant                                                    
#> T/T          24 51.1  68 61.8 1.00             0.22859 196.1
#> C/T-C/C      23 48.9  42 38.2 0.65  0.32  1.31              
#> Recessive                                                   
#> T/T-C/T      45 95.7 100 90.9 1.00             0.28494 196.4
#> C/C           2  4.3  10  9.1 2.22  0.46 10.70              
#> Overdominant                                                
#> T/T-C/C      26 55.3  78 70.9 1.00             0.07188 194.3
#> C/T          21 44.7  32 29.1 0.52  0.25  1.06              
#> log-Additive                                                
#> 0,1,2        47 29.9 110 70.1 0.87  0.51  1.49 0.60861 197.3

# quantitative trait, crude analysis
association(log(protein)~snp10001,data=datSNP)
#> 
#> SNP: snp10001  adjusted by: 
#>                n     me      se     dif   lower    upper   p-value   AIC
#> Codominant                                                              
#> T/T           92 10.637 0.05630  0.0000                  0.0005068 316.1
#> C/T           53 10.351 0.10612 -0.2858 -0.5061 -0.06549                
#> C/C           12  9.929 0.24389 -0.7073 -1.0994 -0.31526                
#> Dominant                                                                
#> T/T           92 10.637 0.05630  0.0000                  0.0008307 318.2
#> C/T-C/C       65 10.273 0.09882 -0.3636 -0.5727 -0.15456                
#> Recessive                                                               
#> T/T-C/T      145 10.532 0.05375  0.0000                  0.0028978 320.6
#> C/C           12  9.929 0.24389 -0.6029 -0.9933 -0.21246                
#> Overdominant                                                            
#> T/T-C/C      104 10.555 0.06087  0.0000                  0.0752824 326.4
#> C/T           53 10.351 0.10612 -0.2042 -0.4276  0.01929                
#> log-Additive                                                            
#> 0,1,2                           -0.3247 -0.4849 -0.16442 0.0001093 314.4
# quantitative trait, adjusted by sex
association(log(protein)~snp10001+sex,data=datSNP)
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> 
#> SNP: snp10001  adjusted by: sex 
#>                n     me      se     dif   lower    upper   p-value   AIC
#> Codominant                                                              
#> T/T           92 10.637 0.05630  0.0000                  2.772e-04 316.0
#> C/T           53 10.351 0.10612 -0.2970 -0.5171 -0.07694                
#> C/C           12  9.929 0.24389 -0.7422 -1.1358 -0.34861                
#> Dominant                                                                
#> T/T           92 10.637 0.05630  0.0000                  5.555e-04 318.6
#> C/T-C/C       65 10.273 0.09882 -0.3774 -0.5871 -0.16762                
#> Recessive                                                               
#> T/T-C/T      145 10.532 0.05375  0.0000                  1.976e-03 321.0
#> C/C           12  9.929 0.24389 -0.6297 -1.0218 -0.23761                
#> Overdominant                                                            
#> T/T-C/C      104 10.555 0.06087  0.0000                  6.889e-02 327.4
#> C/T           53 10.351 0.10612 -0.2091 -0.4329  0.01461                
#> log-Additive                                                            
#> 0,1,2                           -0.3393 -0.5003 -0.17829 5.918e-05 314.3

#
# Interaction analysis
#

# Interaction SNP and factor
association(log(protein)~snp10001*sex+blood.pre, data=datSNP, 
          model="codominant")
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> 
#>       SNP: snp10001  adjusted by: blood.pre 
#>  Interaction 
#> ---------------------
#>         Male             dif   lower   upper    Female              dif   lower
#> T/T 40 10.65 0.08021  0.0000      NA      NA 52 10.626 0.07883 -0.02559 -0.2941
#> C/T 27 10.52 0.10292 -0.1298 -0.4481  0.1886 26 10.176 0.18415 -0.47199 -0.7943
#> C/C  8 10.01 0.34582 -0.6371 -1.1316 -0.1426  4  9.761 0.28635 -0.88728 -1.5571
#>       upper
#> T/T  0.2429
#> C/T -0.1497
#> C/C -0.2174
#> 
#> p interaction: 0.36299 
#> 
#>  sex within snp10001 
#> ---------------------
#> T/T 
#>         n    me      se      dif   lower  upper
#> Male   40 10.65 0.08021  0.00000      NA     NA
#> Female 52 10.63 0.07883 -0.02559 -0.2941 0.2429
#> 
#> C/T 
#>         n    me     se     dif   lower   upper
#> Male   27 10.52 0.1029  0.0000      NA      NA
#> Female 26 10.18 0.1841 -0.3422 -0.6931 0.00864
#> 
#> C/C 
#>        n     me     se     dif  lower  upper
#> Male   8 10.014 0.3458  0.0000     NA     NA
#> Female 4  9.761 0.2863 -0.2502 -1.032 0.5319
#> 
#> p trend: 0.26751 
#> 
#>  snp10001 within sex 
#> ---------------------
#> Male 
#>      n    me      se     dif   lower   upper
#> T/T 40 10.65 0.08021  0.0000      NA      NA
#> C/T 27 10.52 0.10292 -0.1298 -0.4481  0.1886
#> C/C  8 10.01 0.34582 -0.6371 -1.1316 -0.1426
#> 
#> Female 
#>      n     me      se     dif   lower   upper
#> T/T 52 10.626 0.07883  0.0000      NA      NA
#> C/T 26 10.176 0.18415 -0.4464 -0.7538 -0.1390
#> C/C  4  9.761 0.28635 -0.8617 -1.5245 -0.1989
#> 
#> p trend: 0.36299 

# Interaction SNP and SNP (codominant and codominant)
association(log(protein)~snp10001*factor(snp10002)+blood.pre, 
          data=datSNP, model="codominant")
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> 
#>       SNP: snp10001  adjusted by: blood.pre 
#>  Interaction 
#> ---------------------
#>           C/C             dif   lower   upper      A/C              dif   lower
#> T/T 38 10.591 0.1004  0.00000      NA      NA 49 10.66 0.07051  0.06403 -0.2133
#> C/T 24 10.502 0.1622 -0.08375 -0.4189  0.2514 29 10.23 0.13823 -0.36600 -0.6821
#> C/C 12  9.929 0.2439 -0.66157 -1.0860 -0.2371  0  0.00 0.00000       NA      NA
#>        upper n    me  A/A    dif   lower  upper
#> T/T  0.34134 5 10.78 0.15 0.1799 -0.4327 0.7925
#> C/T -0.04993 0  0.00 0.00     NA      NA     NA
#> C/C       NA 0  0.00 0.00     NA      NA     NA
#> 
#> p interaction: 0.13332 
#> 
#>  factor(snp10002) within snp10001 
#> ---------------------
#> T/T 
#>      n    me      se     dif   lower  upper
#> C/C 38 10.59 0.10044 0.00000      NA     NA
#> A/C 49 10.66 0.07051 0.06403 -0.2133 0.3413
#> A/A  5 10.78 0.14998 0.17988 -0.4327 0.7925
#> 
#> C/T 
#>      n    me     se     dif   lower   upper
#> C/C 24 10.50 0.1622  0.0000      NA      NA
#> A/C 29 10.23 0.1382 -0.2823 -0.6373 0.07276
#> A/A  0  0.00 0.0000      NA      NA      NA
#> 
#> C/C 
#>      n    me     se dif lower upper
#> C/C 12 9.929 0.2439   0    NA    NA
#> A/C  0 0.000 0.0000  NA    NA    NA
#> A/A  0 0.000 0.0000  NA    NA    NA
#> 
#> p trend: 0.41739 
#> 
#>  snp10001 within factor(snp10002) 
#> ---------------------
#> C/C 
#>      n     me     se      dif   lower   upper
#> T/T 38 10.591 0.1004  0.00000      NA      NA
#> C/T 24 10.502 0.1622 -0.08375 -0.4189  0.2514
#> C/C 12  9.929 0.2439 -0.66157 -1.0860 -0.2371
#> 
#> A/C 
#>      n    me      se   dif   lower   upper
#> T/T 49 10.66 0.07051  0.00      NA      NA
#> C/T 29 10.23 0.13823 -0.43 -0.7304 -0.1296
#> C/C  0  0.00 0.00000    NA      NA      NA
#> 
#> A/A 
#>     n    me   se dif lower upper
#> T/T 5 10.78 0.15   0    NA    NA
#> C/T 0  0.00 0.00  NA    NA    NA
#> C/C 0  0.00 0.00  NA    NA    NA
#> 
#> p trend: 0.097576 

# Interaction SNP and SNP (dominant and recessive)
association(log(protein)~snp10001*factor(recessive(snp100019))+blood.pre, 
          data=datSNP, model="dominant")
#> Warning: 'varlist' has changed (from nvar=2) to new 3 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=2) to new 4 after EncodeVars() -- should no longer happen!
#> Warning: 'varlist' has changed (from nvar=3) to new 4 after EncodeVars() -- should no longer happen!
#> 
#>       SNP: snp10001  adjusted by: blood.pre 
#>  Interaction 
#> ---------------------
#>        G/G-C/G             dif   lower    upper      C/C             dif
#> T/T 60  10.650 0.06314  0.0000      NA       NA 32 10.61 0.1117 -0.03823
#> C/T 53  10.351 0.10612 -0.2961 -0.5393 -0.05294  0  0.00 0.0000       NA
#> C/C 12   9.929 0.24389 -0.7191 -1.1257 -0.31256  0  0.00 0.0000       NA
#>       lower  upper
#> T/T -0.3196 0.2432
#> C/T      NA     NA
#> C/C      NA     NA
#> 
#> p interaction: NA 
#> 
#>  factor(recessive(snp100019)) within snp10001 
#> ---------------------
#> T/T 
#>          n    me      se      dif   lower  upper
#> G/G-C/G 60 10.65 0.06314  0.00000      NA     NA
#> C/C     32 10.61 0.11174 -0.03823 -0.3196 0.2432
#> 
#> C/T 
#>          n    me     se dif lower upper
#> G/G-C/G 53 10.35 0.1061   0    NA    NA
#> C/C      0  0.00 0.0000  NA    NA    NA
#> 
#> C/C 
#>          n    me     se dif lower upper
#> G/G-C/G 12 9.929 0.2439   0    NA    NA
#> C/C      0 0.000 0.0000  NA    NA    NA
#> 
#> p trend: NA 
#> 
#>  snp10001 within factor(recessive(snp100019)) 
#> ---------------------
#> G/G-C/G 
#>      n     me      se     dif   lower    upper
#> T/T 60 10.650 0.06314  0.0000      NA       NA
#> C/T 53 10.351 0.10612 -0.2961 -0.5393 -0.05294
#> C/C 12  9.929 0.24389 -0.7191 -1.1257 -0.31256
#> 
#> C/C 
#>      n    me     se dif lower upper
#> T/T 32 10.61 0.1117   0    NA    NA
#> C/T  0  0.00 0.0000  NA    NA    NA
#> C/C  0  0.00 0.0000  NA    NA    NA
#> 
#> p trend: NA