Collapsing (or recoding) genotypes into different categories (generally two) depending on a given genetic mode of inheritance

codominant function recodifies a variable having genotypes depending on the allelic frequency in descending order.

dominant, recessive, and overdominant functions collapse the three categories of a given SNP into two categories as follows: Let 'AA', 'Aa', and 'aa' be the three genotypes. After determining the most frequent allele (let's suppose that 'A' is the major allele) the functions return a vector with to categories as follows. dominant: 'AA' and 'Aa-aa'; recessive: 'AA-Aa' and 'aa'; overdominant: 'AA-aa' vs 'Aa'.

additive function creates a numerical variable, 1, 2, 3 corresponding to the three genotypes sorted out by descending allelic frequency (this model is referred as log-additive).

Usage

codominant(o)
dominant(o)
recessive(o)
overdominant(o)
additive(o)

Arguments

o

categorical covariate having genotypes

Value

A snp object collapsing genotypes into different categories depending on a given genetic mode of inheritance

Examples

data(SNPs)
dominant(snp(SNPs$snp10001,sep=""))
#>   [1] T/T     T/T     T/T     C/T-C/C T/T     T/T     T/T     T/T     C/T-C/C
#>  [10] T/T     C/T-C/C C/T-C/C C/T-C/C T/T     T/T     T/T     C/T-C/C T/T    
#>  [19] T/T     T/T     T/T     C/T-C/C C/T-C/C T/T     T/T     C/T-C/C T/T    
#>  [28] T/T     T/T     C/T-C/C T/T     T/T     C/T-C/C T/T     C/T-C/C C/T-C/C
#>  [37] C/T-C/C C/T-C/C C/T-C/C T/T     T/T     T/T     T/T     C/T-C/C T/T    
#>  [46] C/T-C/C C/T-C/C C/T-C/C C/T-C/C T/T     T/T     C/T-C/C C/T-C/C T/T    
#>  [55] T/T     T/T     T/T     T/T     C/T-C/C T/T     T/T     T/T     T/T    
#>  [64] T/T     T/T     T/T     C/T-C/C T/T     T/T     C/T-C/C C/T-C/C T/T    
#>  [73] T/T     C/T-C/C T/T     T/T     T/T     T/T     T/T     C/T-C/C C/T-C/C
#>  [82] T/T     T/T     T/T     C/T-C/C T/T     T/T     C/T-C/C C/T-C/C C/T-C/C
#>  [91] T/T     T/T     C/T-C/C T/T     T/T     T/T     T/T     T/T     C/T-C/C
#> [100] T/T     C/T-C/C C/T-C/C C/T-C/C T/T     T/T     C/T-C/C C/T-C/C T/T    
#> [109] T/T     C/T-C/C T/T     C/T-C/C C/T-C/C T/T     C/T-C/C C/T-C/C T/T    
#> [118] T/T     C/T-C/C C/T-C/C C/T-C/C T/T     C/T-C/C T/T     T/T     T/T    
#> [127] T/T     T/T     C/T-C/C T/T     C/T-C/C C/T-C/C T/T     C/T-C/C C/T-C/C
#> [136] T/T     C/T-C/C T/T     T/T     C/T-C/C C/T-C/C C/T-C/C T/T     T/T    
#> [145] T/T     T/T     C/T-C/C T/T     T/T     T/T     T/T     C/T-C/C T/T    
#> [154] C/T-C/C C/T-C/C C/T-C/C C/T-C/C
#> Genotypes: T/T C/T-C/C
#> Alleles:  
overdominant(snp(SNPs$snp10001,sep=""))
#>   [1] T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T    
#>  [10] T/T-C/C C/T     T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C
#>  [19] T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C T/T-C/C C/T     T/T-C/C
#>  [28] T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C C/T     C/T    
#>  [37] T/T-C/C C/T     C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C
#>  [46] T/T-C/C C/T     C/T     C/T     T/T-C/C T/T-C/C C/T     C/T     T/T-C/C
#>  [55] T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C
#>  [64] T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     C/T     T/T-C/C
#>  [73] T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     C/T    
#>  [82] T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T    
#>  [91] T/T-C/C T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T    
#> [100] T/T-C/C C/T     C/T     C/T     T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C
#> [109] T/T-C/C C/T     T/T-C/C C/T     C/T     T/T-C/C C/T     C/T     T/T-C/C
#> [118] T/T-C/C C/T     C/T     C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C T/T-C/C
#> [127] T/T-C/C T/T-C/C C/T     T/T-C/C C/T     T/T-C/C T/T-C/C C/T     C/T    
#> [136] T/T-C/C C/T     T/T-C/C T/T-C/C C/T     C/T     C/T     T/T-C/C T/T-C/C
#> [145] T/T-C/C T/T-C/C C/T     T/T-C/C T/T-C/C T/T-C/C T/T-C/C C/T     T/T-C/C
#> [154] C/T     C/T     C/T     C/T    
#> Genotypes: T/T-C/C C/T
#> Alleles: