max-statistic for a 2x3 table

Computes the asymptotic p-value for max-statistic for a 2x3 table

Usage

maxstat(x, ...)

# S3 method for default
maxstat(x, y, ...)

# S3 method for table
maxstat(x, ...)

# S3 method for setupSNP
maxstat(x, y, colSNPs=attr(x,"colSNPs"), ...)

# S3 method for matrix
maxstat(x, ...)

Arguments

x

a numeric matrix with 2 rows (cases/controls) and 3 colums (genotypes) or a vector with case/control status or an object of class 'setupSNP'.

y

an optional numeric vector containing the information for a given SNP. In this case 'x' argument must contain a vector indicarting case/control status. If 'x' argument is an object of class 'setupSNP' this argument migth be the name of the variable containing case/control information.

colSNPs

a vector indicating which columns contain those SNPs to compute max-statistic. By default max-statistic is computed for those SNPs specified when the object of class 'setupSNP' was created.

...

further arguments to be passed to or from methods.

References

Gonzalez JR, Carrasco JL, Dudbridge F, Armengol L, Estivill X, Moreno V. Maximizing association statistics over genetic models (2007). Submitted

Sladek R, Rocheleau G, Rung J et al. A genome-wide association study identifies novel risk loci for type 2 diabetes (2007). Nature 445, 881-885

Value

A matrix with the chi-square statistic for dominant, recessive, log-additive and max-statistic and its asymptotic p-value.

See also

Examples

# example from Sladek et al. (2007) for the SNP rs1111875 
 tt<-matrix(c(77,298,310,122,316,231),nrow=2,ncol=3,byrow=TRUE)
 maxstat(tt)
#>      dominant recessive log-additive MAX-statistic   Pr(>z)    
#> [1,]   13.298    16.275       22.019        22.019 3.76e-06 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

 data(SNPs)
 maxstat(SNPs$casco,SNPs$snp10001) 
#>      dominant recessive log-additive MAX-statistic Pr(>z)
#> [1,]    1.209     1.559        0.286         1.559  0.384
 myDat<-setupSNP(SNPs,6:40,sep="")
 maxstat(myDat,casco)
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP only has 2 genotypes
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#> Warning: this SNP is monomorphic
#>           dominant recessive log-additive MAX-statistic Pr(>z)
#> snp10001     1.559     1.209        0.286         1.559  0.384
#> snp10002     0.012     1.479        0.068         1.479  0.411
#> snp10003         -         -            -             -      -
#> snp10004         -         -            -             -      -
#> snp10005     0.013     0.491        0.005         0.491  0.741
#> snp10006         -         -            -             -      -
#> snp10007         -         -            -             -      -
#> snp10008     0.173     0.286        0.017         0.286  0.831
#> snp10009     0.019     1.429        0.053         1.429  0.424
#> snp100010        -         -            -             -      -
#> snp100011    0.258     0.007        0.147         0.258  0.818
#> snp100012    0.022     0.444        0.094         0.444  0.762
#> snp100013    0.620     0.425        0.126         0.620  0.670
#> snp100014    0.588     0.327        0.043         0.588  0.685
#> snp100015    0.005     0.005        0.005         0.005  0.946
#> snp100016        -         -            -             -      -
#> snp100017    0.046     1.379        0.026         1.379  0.437
#> snp100018    0.103     1.429        0.005         1.429  0.425
#> snp100019    1.264     0.370        0.155         1.264  0.451
#> snp100020    0.279     0.286        0.047         0.286  0.832
#> snp100021        -         -            -             -      -
#> snp100022        -         -            -             -      -
#> snp100023    0.017     1.473        0.060         1.473  0.412
#> snp100024    1.012     0.508        0.209         1.012  0.528
#> snp100025        -         -            -             -      -
#> snp100026        -         -            -             -      -
#> snp100027    0.223     1.451        0.004         1.451  0.420
#> snp100028    1.186     0.540        0.080         1.186  0.470
#> snp100029    0.378     0.508        0.025         0.508  0.719
#> snp100030        -         -            -             -      -
#> snp100031        -         -            -             -      -
#> snp100032    0.737     0.540        0.019         0.737  0.619
#> snp100033    0.885     0.664        0.018         0.885  0.564
#> snp100034    0.378     0.508        0.025         0.508  0.719
#> snp100035        -         -            -             -      -