Test for Hardy-Weinberg Equilibrium

Test the null hypothesis that Hardy-Weinberg equilibrium holds in cases, controls and both populations.

print print the information. Number of digits, and significance level can be changed

Usage

tableHWE(x, strata, ...)

Arguments

x

an object of class 'setupSNP'

strata

a factor variable for a stratified analysis

...

optional arguments

Details

This function calculates the HWE test for those variables of class 'snp' in the object x of class 'setupSNP'

Value

A matrix with p values for Hardy-Weinberg Equilibrium

Author

This function is based on an R function which computes an exact SNP test of Hardy-Weinberg Equilibrium written by Wigginton JE, Cutler DJ and Abecasis GR available at http://csg.sph.umich.edu/abecasis/Exact/r_instruct.html

See also

References

Wigginton JE, Cutler DJ and Abecasis GR (2005). A note on exact tests of Hardy-Weinberg equilibrium. Am J Hum Genet 76:887-93

Examples

data(SNPs)
ans<-setupSNP(SNPs,6:40,sep="")
res<-tableHWE(ans)
print(res)
#>           HWE (p value) flag
#> snp10001  0.2816            
#> snp10002  0.0049        <-  
#> snp10003  -                 
#> snp10004  -                 
#> snp10005  0.0080        <-  
#> snp10006  -                 
#> snp10007  -                 
#> snp10008  0.1378            
#> snp10009  0.0028        <-  
#> snp100010 -                 
#> snp100011 0.0191        <-  
#> snp100012 0.0134        <-  
#> snp100013 0.0256        <-  
#> snp100014 1.0000            
#> snp100015 -                 
#> snp100016 -                 
#> snp100017 0.0005        <-  
#> snp100018 0.0005        <-  
#> snp100019 0.7463            
#> snp100020 0.1254            
#> snp100021 -                 
#> snp100022 -                 
#> snp100023 0.0028        <-  
#> snp100024 0.0922            
#> snp100025 -                 
#> snp100026 -                 
#> snp100027 0.0009        <-  
#> snp100028 0.4197            
#> snp100029 0.0487        <-  
#> snp100030 -                 
#> snp100031 -                 
#> snp100032 0.2589            
#> snp100033 0.3264            
#> snp100034 0.0487        <-  
#> snp100035 -                 
#change the significance level showed in the flag column
print(res,sig=0.001)
#>           HWE (p value) flag
#> snp10001  0.2816            
#> snp10002  0.0049            
#> snp10003  -                 
#> snp10004  -                 
#> snp10005  0.0080            
#> snp10006  -                 
#> snp10007  -                 
#> snp10008  0.1378            
#> snp10009  0.0028            
#> snp100010 -                 
#> snp100011 0.0191            
#> snp100012 0.0134            
#> snp100013 0.0256            
#> snp100014 1.0000            
#> snp100015 -                 
#> snp100016 -                 
#> snp100017 0.0005        <-  
#> snp100018 0.0005        <-  
#> snp100019 0.7463            
#> snp100020 0.1254            
#> snp100021 -                 
#> snp100022 -                 
#> snp100023 0.0028            
#> snp100024 0.0922            
#> snp100025 -                 
#> snp100026 -                 
#> snp100027 0.0009        <-  
#> snp100028 0.4197            
#> snp100029 0.0487            
#> snp100030 -                 
#> snp100031 -                 
#> snp100032 0.2589            
#> snp100033 0.3264            
#> snp100034 0.0487            
#> snp100035 -                 

#stratified analysis
res<-tableHWE(ans,ans$sex)
print(res)
#>           all.groups   Male Female
#> snp10001      0.2816 0.3941 0.7388
#> snp10002      0.0049 0.1660 0.0075
#> snp10003           -      -      -
#> snp10004           -      -      -
#> snp10005      0.0080 0.2755 0.0257
#> snp10006           -      -      -
#> snp10007           -      -      -
#> snp10008      0.1378 0.5078 0.2342
#> snp10009      0.0028 0.0992 0.0075
#> snp100010          -      -      -
#> snp100011     0.0191 1.0000 0.0184
#> snp100012     0.0134 0.2761 0.0255
#> snp100013     0.0256 0.1206 0.2051
#> snp100014     1.0000 0.8101 0.6456
#> snp100015          -      -      -
#> snp100016          -      -      -
#> snp100017     0.0005 0.0304 0.0068
#> snp100018     0.0005 0.0304 0.0066
#> snp100019     0.7463 1.0000 0.5012
#> snp100020     0.1254 0.5078 0.2141
#> snp100021          -      -      -
#> snp100022          -      -      -
#> snp100023     0.0028 0.0972 0.0123
#> snp100024     0.0922 0.1551 0.5197
#> snp100025          -      -      -
#> snp100026          -      -      -
#> snp100027     0.0009 0.0304 0.0123
#> snp100028     0.4197 1.0000 0.2619
#> snp100029     0.0487 0.0772 0.5065
#> snp100030          -      -      -
#> snp100031          -      -      -
#> snp100032     0.2589 0.8170 0.1834
#> snp100033     0.3264 0.8139 0.2619
#> snp100034     0.0487 0.0772 0.5065
#> snp100035          -      -      -